+++ changelog.rst
@@ -5,6 +5,10 @@ Moldy Bread Change Log
 Not Yet Tagged for Release  / Upcoming
 ======================================
 
+**New Operations**:
+
+* Download book pages and name semantically as book_pageNumber.extension.  See docs for more details.
+
 **New Fedora Public Methods and Other Procs**:
 
 * FedoraRequest.download_page_with_book_relatioship():  Downloads all matching datastreams related to a page with naming based on the book and page number.